Revert "Move plot_bar_chart method to StatsCommandsCog"
This reverts commit b9b5b947a24a6a549d0653a97786d724f93b6aa3.
